Implement DigitInfo class
In digits method, return digits of the number in a list. Return the sum
of digits in sum method. In sum_of_squares method, return sum of squares
of digits. In square_of_sum return square of sum of digits. Provide a
method to check if the number is palindrome. Also provide a method to
check if the number is armstrong.